requirements.txt
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.0
idna==3.3
oauthlib==3.2.0
pip==21.2.4
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
setuptools==58.0.4
tweepy==4.5.0
urllib3==1.26.8
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
